+++ lib/Applications.ps1
@@ -1,5 +1,15 @@
-'core', 'json', 'Helpers', 'manifest', 'Versions' | ForEach-Object {
-    . (Join-Path $PSScriptRoot "$_.ps1")
+@(
+    @('core', 'Test-ScoopDebugEnabled'),
+    @('Helpers', 'New-IssuePrompt'),
+    @('json', 'ConvertToPrettyJson'),
+    @('manifest', 'Resolve-ManifestInformation'),
+    @('depends', 'script_deps'),
+    @('Versions', 'Clear-InstalledVersion')
+) | ForEach-Object {
+    if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
+        Write-Verbose "Import of lib '$($_[0])' initiated from '$PSCommandPath'"
+        . (Join-Path $PSScriptRoot "$($_[0]).ps1")
+    }
 }
 
 #region Application instalaltion info file
@@ -157,3 +167,85 @@ function Set-InstalledApplicationInformationPropertyValue {
     }
 }
 #endregion Application instalaltion info file
+
+function app_status($app, $global) {
+    $status = @{ }
+    $status.installed = (installed $app $global)
+    $status.version = Select-CurrentVersion -AppName $app -Global:$global
+    $status.latest_version = $status.version
+
+    $install_info = install_info $app $status.version $global
+
+    $status.failed = (!$install_info -or !$status.version)
+    $status.hold = ($install_info.hold -eq $true)
+
+    $manifest = manifest $app $install_info.bucket $install_info.url
+    $status.bucket = $install_info.bucket
+    $status.removed = (!$manifest)
+    if ($manifest.version) {
+        $status.latest_version = $manifest.version
+    }
+
+    $status.outdated = $false
+    if ($status.version -and $status.latest_version) {
+        $status.outdated = (Compare-Version -ReferenceVersion $status.version -DifferenceVersion $status.latest_version) -ne 0
+    }
+
+    $status.missing_deps = @()
+    $deps = @(runtime_deps $manifest) | Where-Object {
+        $app, $bucket, $null = parse_app $_
+        return !(installed $app)
+    }
+
+    if ($deps) { $status.missing_deps += , $deps }
+
+    return $status
+}
+
+function Confirm-InstallationStatus {
+    <#
+    .SYNOPSIS
+        Get status of specific applications.
+        Returns array of 3 item arrays (appliation name, globally installed, bucket name)
+    .PARAMETER Apps
+        Specifies the array of applications to be evalueated.
+    .PARAMETER Global
+        Specifies to check globally installed applications.
+    #>
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)]
+        [String[]] $Apps,
+        [Switch] $Global
+    )
+    $Global | Out-Null # PowerShell/PSScriptAnalyzer#1472
+    $installed = @()
+
+    $Apps | Select-Object -Unique | Where-Object -Property 'Name' -NE -Value 'scoop' | ForEach-Object {
+        # TODO: Adopt Resolve-ManifestInformation
+        # Should not be needed to resolve, as it will contain only valid installed applications
+        $app, $null, $null = parse_app $_
+        $buc = (app_status $app $Global).bucket
+        if ($Global) {
+            if (installed $app $true) {
+                $installed += , @($app, $true, $buc)
+            } elseif (installed $app $false) {
+                Write-UserMessage -Message "'$app' isn't installed globally, but it is installed for your account." -Err
+                Write-UserMessage -Message 'Try again without the --global (or -g) flag instead.' -Warning
+            } else {
+                Write-UserMessage -Message "'$app' isn't installed." -Err
+            }
+        } else {
+            if (installed $app $false) {
+                $installed += , @($app, $false, $buc)
+            } elseif (installed $app $true) {
+                Write-UserMessage -Message "'$app' isn't installed for your account, but it is installed globally." -Err
+                Write-UserMessage -Message 'Try again with the --global (or -g) flag instead.' -Warning
+            } else {
+                Write-UserMessage -Message "'$app' isn't installed." -Err
+            }
+        }
+    }
+
+    return , $installed
+}