+++ lib/core.ps1
@@ -1,5 +1,11 @@
-'Helpers', 'commands' | ForEach-Object {
-    . (Join-Path $PSScriptRoot "$_.ps1")
+@(
+    @('Helpers', 'New-IssuePrompt'),
+    @('Helpers', 'New-IssuePrompt')
+) | ForEach-Object {
+    if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
+        Write-Verbose "Import of lib '$($_[0])' initiated from '$PSCommandPath'"
+        . (Join-Path $PSScriptRoot "$($_[0]).ps1")
+    }
 }
 
 # Such format is need to prevent automatic conversion of JSON date https://github.com/Ash258/Scoop-Core/issues/26
@@ -445,42 +451,6 @@ function Test-Aria2Enabled {
     process { return (Test-HelperInstalled -Helper 'Aria2') -and (get_config 'aria2-enabled' $true) }
 }
 
-function app_status($app, $global) {
-    $status = @{ }
-    $status.installed = (installed $app $global)
-    $status.version = Select-CurrentVersion -AppName $app -Global:$global
-    $status.latest_version = $status.version
-
-    $install_info = install_info $app $status.version $global
-
-    $status.failed = (!$install_info -or !$status.version)
-    $status.hold = ($install_info.hold -eq $true)
-
-    $manifest = manifest $app $install_info.bucket $install_info.url
-    $status.bucket = $install_info.bucket
-    $status.removed = (!$manifest)
-    if ($manifest.version) {
-        $status.latest_version = $manifest.version
-    }
-
-    $status.outdated = $false
-    if ($status.version -and $status.latest_version) {
-        $status.outdated = (Compare-Version -ReferenceVersion $status.version -DifferenceVersion $status.latest_version) -ne 0
-    }
-
-    $status.missing_deps = @()
-    # TODO: Eliminate
-    . (Join-Path $PSScriptRoot 'depends.ps1')
-    $deps = @(runtime_deps $manifest) | Where-Object {
-        $app, $bucket, $null = parse_app $_
-        return !(installed $app)
-    }
-
-    if ($deps) { $status.missing_deps += , $deps }
-
-    return $status
-}
-
 # paths
 function fname($path) { return Split-Path $path -Leaf }
 function strip_ext($fname) { return $fname -replace '\.[^\.]*$' }
@@ -842,54 +812,6 @@ function ensure_architecture($architecture_opt) {
     }
 }
 
-function Confirm-InstallationStatus {
-    <#
-    .SYNOPSIS
-        Get status of specific applications.
-        Returns array of 3 item arrays (appliation name, globally installed, bucket name)
-    .PARAMETER Apps
-        Specifies the array of applications to be evalueated.
-    .PARAMETER Global
-        Specifies to check globally installed applications.
-    #>
-    [CmdletBinding()]
-    param(
-        [Parameter(Mandatory)]
-        [String[]] $Apps,
-        [Switch] $Global
-    )
-    $Global | Out-Null # PowerShell/PSScriptAnalyzer#1472
-    $installed = @()
-
-    $Apps | Select-Object -Unique | Where-Object -Property 'Name' -NE -Value 'scoop' | ForEach-Object {
-        # TODO: Adopt Resolve-ManifestInformation
-        # Should not be needed to resolve, as it will contain only valid installed applications
-        $app, $null, $null = parse_app $_
-        $buc = (app_status $app $Global).bucket
-        if ($Global) {
-            if (installed $app $true) {
-                $installed += , @($app, $true, $buc)
-            } elseif (installed $app $false) {
-                Write-UserMessage -Message "'$app' isn't installed globally, but it is installed for your account." -Err
-                Write-UserMessage -Message 'Try again without the --global (or -g) flag instead.' -Warning
-            } else {
-                Write-UserMessage -Message "'$app' isn't installed." -Err
-            }
-        } else {
-            if (installed $app $false) {
-                $installed += , @($app, $false, $buc)
-            } elseif (installed $app $true) {
-                Write-UserMessage -Message "'$app' isn't installed for your account, but it is installed globally." -Err
-                Write-UserMessage -Message 'Try again with the --global (or -g) flag instead.' -Warning
-            } else {
-                Write-UserMessage -Message "'$app' isn't installed." -Err
-            }
-        }
-    }
-
-    return , $installed
-}
-
 function strip_path($orig_path, $dir) {
     if ($null -eq $orig_path) { $orig_path = '' }
     $stripped = [string]::join(';', @( $orig_path.split(';') | Where-Object { $_ -and $_ -ne $dir } ))
@@ -1117,12 +1039,19 @@ $SCOOP_ROOT_DIRECTORY = $env:SCOOP, "$env:USERPROFILE\scoop" | Where-Object { -n
 # Scoop global apps directory
 $SCOOP_GLOBAL_ROOT_DIRECTORY = $env:SCOOP_GLOBAL, "$env:ProgramData\scoop" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
 
+# Directory for local buckets
+$SCOOP_BUCKETS_DIRECTORY = Join-Path $SCOOP_ROOT_DIRECTORY 'buckets'
+
 # Scoop cache directory
 # Note: Setting the SCOOP_CACHE environment variable to use a shared directory
 #       is experimental and untested. There may be concurrency issues when
 #       multiple users write and access cached files at the same time.
 #       Use at your own risk.
 $SCOOP_CACHE_DIRECTORY = $env:SCOOP_CACHE, "$SCOOP_ROOT_DIRECTORY\cache" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
+
+# Scoop directory for powershell modules installtation
+$SCOOP_MODULE_DIRECTORY = Join-Path $SCOOP_ROOT_DIRECTORY 'modules'
+$SCOOP_GLOBAL_MODULE_DIRECTORY = Join-Path $SCOOP_GLOBAL_ROOT_DIRECTORY 'modules'
 
 # Directory for downloaded manifests (mainly)
 $SHOVEL_GENERAL_MANIFESTS_DIRECTORY = Join-Path $SCOOP_ROOT_DIRECTORY 'manifests'
@@ -1138,6 +1067,8 @@ $globaldir = $SCOOP_GLOBAL_ROOT_DIRECTORY
 $cachedir = $SCOOP_CACHE_DIRECTORY
 $scoopConfig = $SCOOP_CONFIGURATION
 $configFile = $SCOOP_CONFIGURATION_FILE
+$modulesdir = $SCOOP_MODULE_DIRECTORY
+$bucketsdir = $SCOOP_BUCKETS_DIRECTORY
 
 # Do not use the new native command parsing PowerShell/PowerShell#15239, Ash258/Scoop-Core#142
 $PSNativeCommandArgumentPassing = 'Legacy'