+++ lib/decompress.ps1
@@ -1,6 +1,11 @@
-# TODO: Core import is messing up with download progress
-'Helpers' | ForEach-Object { #, 'core' | ForEach-Object {
-    . (Join-Path $PSScriptRoot "$_.ps1")
+@(
+    @('core', 'Test-ScoopDebugEnabled'),
+    @('Helpers', 'New-IssuePrompt')
+) | ForEach-Object {
+    if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
+        Write-Verbose "Import of lib '$($_[0])' initiated from '$PSCommandPath'"
+        . (Join-Path $PSScriptRoot "$($_[0]).ps1")
+    }
 }
 
 #region helpers
@@ -54,7 +59,7 @@ function Test-LessmsiRequirement {
 
     if ($null -eq $URL) { return $false }
 
-    if (get_config 'MSIEXTRACT_USE_LESSMSI' $false) {
+    if (get_config 'MSIEXTRACT_USE_LESSMSI' $true) {
         return ($URL | Where-Object { $_ -match '\.msi$' }).Count -gt 0
     } else {
         return $false
@@ -82,7 +87,7 @@ function Test-ZstdRequirement {
 }
 
 function _decompressErrorPrompt($path, $log) {
-    return @("Decompress error|-Failed to extract files from $path.", "Log file:", "  $(friendly_path $log)") -join "`n"
+    return @("Decompress error|-Failed to extract files from $path.", 'Log file:', "  $(friendly_path $log)") -join "`n"
 }
 #endregion helpers
 
@@ -213,7 +218,7 @@ function Expand-MsiArchive {
             $DestinationPath = Join-Path $DestinationPath '_tmp'
         }
 
-        if ((get_config 'MSIEXTRACT_USE_LESSMSI' $false)) {
+        if ((get_config 'MSIEXTRACT_USE_LESSMSI' $true)) {
             $msiPath = Get-HelperPath -Helper 'Lessmsi'
             $argList = @('x', "`"$Path`"", "`"$DestinationPath\\`"")
         } else {