+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [0.6.5](https://github.com/Ash258/Scoop-Core/milestone/5)
 
+- **scoop-config**:
+    - Consider lessmsi as default for msi extraction
+    - Add config option `dbgBypassArmCheck`
+        - Installation will not fail when installing arm64 version on x86.
+        - For debugging purpose, to simulate the installation of arm64 version
+- Internal modules/libs are loaded only when not loaded already
 - **scoop-install**: Fix independent installations
 - Initial support for manipulation with `arm64` based manifests
 - **Update**: Application will not be updated, if manifest does not support architecture used for installation
@@ -15,6 +21,7 @@
     - Support `githubToken` config and `GITHUB_TOKEN` environment variable for Github API calls
 - **scoop-install**, **scoop-update**: Report failed installations/updates at the end of execution
 - **Schema**:
+    - Fix validation for curly bracket variables in autoupdate
     - Initial support for `arm64` architecture
     - Allow `$schema` property
 - **CI**: